ENH: optimisation of routing and knn
Optimisations are done via LookupTables and Tensors.
Routing becomes oblivious transfer (OT) of next node to visit.
By repeating the request with the new node as origin, the client can obtain the full path.

Nearest neighbours is done via swap sort to compute the equivalent of argpartition of an array of distances.